Project 4 - Facial Expression Recognition/requirements.txt
pandas
numpy
scikit-learn
pillow
